lib: add symbols instead of methods to hide impl details
