+++ lib/internal/http2/core.js
@@ -106,7 +106,11 @@ const {
   validateTimerDuration,
   refreshFnSymbol
 } = require('internal/timers');
-const StreamBase = require('internal/streams/stream_base');
+const {
+  kCreateWriteWrap,
+  kWriteGeneric,
+  StreamBase
+} = require('internal/stream_base');
 
 const { ShutdownWrap, WriteWrap } = process.binding('stream_wrap');
 const { constants, nameForErrorCode } = binding;
@@ -1653,8 +1657,8 @@ class Http2Stream extends Duplex {
     const req = new WriteWrap();
     req.stream = this[kID];
     req.callback = cb;
-    this.createWriteWrap(req, handle, afterDoStreamWrite);
-    this.writeGeneric(req, data, encoding, cb);
+    this[kCreateWriteWrap](req, handle, afterDoStreamWrite);
+    this[kWriteGeneric](req, data, encoding, cb);
 
     trackWriteState(this, req.bytes);
   }
@@ -1883,7 +1887,7 @@ class Http2Stream extends Duplex {
     }
   }
 }
-util._extend(Http2Stream.prototype, StreamBase);
+StreamBase.apply(Http2Stream.prototype);
 
 function processHeaders(headers) {
   assertIsObject(headers, 'headers');
