+++ lib/net.js
@@ -52,7 +52,12 @@ const {
   defaultTriggerAsyncIdScope,
   symbols: { async_id_symbol }
 } = require('internal/async_hooks');
-const StreamBase = require('internal/streams/stream_base');
+const {
+  kCreateWriteWrap,
+  kWritevGeneric,
+  kWriteGeneric,
+  StreamBase
+} = require('internal/stream_base');
 const errors = require('internal/errors');
 const {
   ERR_INVALID_ADDRESS_FAMILY,
@@ -309,7 +314,7 @@ function Socket(options) {
   this[BYTES_READ] = 0;
 }
 util.inherits(Socket, stream.Duplex);
-util._extend(Socket.prototype, StreamBase);
+StreamBase.apply(Socket.prototype);
 
 // Refresh existing timeouts.
 Socket.prototype._unrefTimer = function _unrefTimer() {
@@ -745,9 +750,11 @@ Socket.prototype._writeGeneric = function(writev, data, encoding, cb) {
   let ret;
   const req = new WriteWrap();
   const handle = this._handle;
-  this.createWriteWrap(req, handle, afterWrite);
-  if (writev) ret = this.writevGeneric(req, data, cb);
-  else ret = this.writeGeneric(req, data, encoding, cb);
+  this[kCreateWriteWrap](req, handle, afterWrite);
+  if (writev)
+    ret = this[kWritevGeneric](req, data, cb);
+  else
+    ret = this[kWriteGeneric](req, data, encoding, cb);
 
   if (ret) return ret;
