+++ node.gyp
@@ -144,8 +144,8 @@
       'lib/internal/v8_prof_polyfill.js',
       'lib/internal/v8_prof_processor.js',
       'lib/internal/vm/Module.js',
+      'lib/internal/stream_base.js',
       'lib/internal/streams/lazy_transform.js',
-      'lib/internal/streams/stream_base.js',
       'lib/internal/streams/async_iterator.js',
       'lib/internal/streams/BufferList.js',
       'lib/internal/streams/duplex_base.js',
